requirements.txt
streamlit
pillow
pydrive2
